added badge helper and updated code
